+++ docs/CHANGELOG.md
@@ -1,4 +1,12 @@
 <!-- PLAYBOOK:CHANGELOG_RELEASE_NOTES_START -->
+## 0.33.0 - 2026-03-26
+- Recommended bump: minor
+- @fawxzzy/playbook: 0.32.0 -> 0.33.0 (playbook-installable-workspace)
+- @fawxzzy/playbook-cli: 0.32.0 -> 0.33.0 (playbook-installable-workspace)
+- @zachariahredfield/playbook-core: 0.32.0 -> 0.33.0 (playbook-installable-workspace)
+- @zachariahredfield/playbook-engine: 0.32.0 -> 0.33.0 (playbook-installable-workspace)
+- @zachariahredfield/playbook-node: 0.32.0 -> 0.33.0 (playbook-installable-workspace)
+
 ## 0.32.0 - 2026-03-26
 - Recommended bump: minor
 - @fawxzzy/playbook: 0.31.0 -> 0.32.0 (playbook-installable-workspace)